[220602][FE] Refactor: 유저, 관리자 페이지 리펙토링

See merge request sw_track/class_02_seoul/web_project/team19/shopping-mall!67

+++ src/views/account/account.js
@@ -1,29 +1,37 @@
 import * as Api from '../api.js';
 import { loginUser, logoutUser } from '../useful-functions.js';
 
-const deleteButton = document.querySelector('#deleteButton');
+const userInformation = document.querySelector('#userInformation');
 
-(() => {
-  renderUserInfo();
+init();
+
+function init() {
   loginUser();
   logoutUser();
+  renderUserInfo();
   addAllEvents();
-})();
+}
 
 function addAllEvents() {
+  const deleteButton = document.querySelector('#deleteButton');
   deleteButton.addEventListener('click', deleteUserInfo);
 }
 
 async function renderUserInfo() {
-  const user = await Api.get('/api/user/info');
-  addUserInfo(user);
+  try {
+    const user = await Api.get('/api/user/info');
+    addUserInfo(user);
+  } catch (error) {
+    alert('회원 정보를 불러오는 데 실패했습니다. 다시 한번 시도해 주세요.');
+    window.location.reload();
+  }
 }
 
 function addUserInfo(user) {
-  const userName = document.querySelector('#userName');
-  const userEmail = document.querySelector('#userEmail');
-  const userPhoneNumber = document.querySelector('#userPhoneNumber');
-  const userAddress = document.querySelector('#userAddress');
+  const userName = userInformation.querySelector('#userName');
+  const userEmail = userInformation.querySelector('#userEmail');
+  const userPhoneNumber = userInformation.querySelector('#userPhoneNumber');
+  const userAddress = userInformation.querySelector('#userAddress');
 
   userName.innerText = user.fullName;
   userEmail.innerText = user.email;
@@ -34,12 +42,19 @@ function addUserInfo(user) {
 }
 
 async function deleteUserInfo() {
-  try {
-    const result = await Api.delete('/api/user/delete');
-    alert(`${result.fullName}님의 계정이 탈퇴되었습니다.`);
-    sessionStorage.removeItem('token');
-    window.location.href = '/';
-  } catch (err) {
-    alert('회원 탈퇴처리가 정상적으로 진행되지 않았습니다. 다시 시도해주세요.');
+  const deleteFlag = confirm(
+    '탈퇴시 삭제된 정보는 다시 복구되지 않습니다. 탈퇴 하시겠습니까?'
+  );
+  if (deleteFlag) {
+    try {
+      const result = await Api.delete('/api/user/delete');
+      alert(`${result.fullName}님의 계정이 탈퇴되었습니다.`);
+      sessionStorage.removeItem('token');
+      window.location.href = '/';
+    } catch (err) {
+      alert(
+        '회원 탈퇴처리가 정상적으로 진행되지 않았습니다. 다시 시도해주세요.'
+      );
+    }
   }
 }
